add tests

+++ hw4/hw4.1/Calculator/Calculator/Nodes/INode.cs
@@ -13,6 +13,6 @@
         /// <summary>
         /// Calculates this Node.
         /// </summary>
-        public int Calculate();
+        public double Calculate();
     }
 }
