+++ hw4/hw4.1/Calculator/Calculator/Nodes/Number.cs
@@ -9,7 +9,7 @@
         private readonly int _value;
         public Number(int value) => this._value = value;
 
-        public int Calculate() => _value;
+        public double Calculate() => _value;
 
         public string Print() => _value.ToString();
     }
