+++ hw4/hw4.1/Calculator/Calculator/Solver.cs
@@ -55,6 +55,6 @@ namespace Calculator
         /// <summary>
         /// Calculates this expression.
         /// </summary>
-        public int Calculate() => Expression.Calculate();
+        public double Calculate() => Expression.Calculate();
     }
 }
